requirements.txt
fastapi
uvicorn
requests
apscheduler
python-dotenv
